refactor: devide section component into multiple components
