+++ src/FavoriteProductSection.js
@@ -1,10 +1,13 @@
-import ProductList from './ProductList';
+import React from 'react';
+import ProductSectionHeader from './ProductSectionHeader.js';
+import ProductList from './ProductList.js';
 import './css/product-section.css';
 
 export default function FavoriteProductSection({ items }) {
+  const HEADER_TEXT = '베스트 상품';
   return (
     <section className="product-section product-section__favorite">
-      <h2 className="product-section__header-text">베스트 상품</h2>
+      <ProductSectionHeader text={HEADER_TEXT} />
       <ProductList className="product-list__favorite" items={items} />
     </section>
   );
